Add files via upload
Welcome to the repository for the Sentiment Analysis on E-Commerce Reviews using LSTM networks. This project aims to classify the sentiments of customer reviews as positive, negative, or neutral by leveraging deep learning techniques. The dataset used is the Amazon Fine Food Reviews from Kaggle, consisting of 568,000 reviews. This repository includes all the necessary code, datasets, and instructions to preprocess data, train the model, and evaluate its performance.